Fully added localization support (currently translated to english and russian); Project directory refactoring (ANNA -> anna, etc.); Code cleanup and minor changes.
