+++ src/anna/network/neurons/BasicNeuron.java
@@ -1,0 +1,27 @@
+package anna.network.neurons;
+
+import anna.math.ActivationFunctions;
+import anna.network.Hyperparameters;
+import anna.network.NetworkStructure;
+
+public class BasicNeuron extends Neuron {
+    public BasicNeuron(int id, ActivationFunctions.types activationFunction, NetworkStructure.neuronTypes type) {
+        super(id, activationFunction, type);
+    }
+
+    @Override
+    protected double calculateRawOutput(double[] inputs, double[] weights) {
+        double rawOutput = 0;
+
+        //Multiply by weights and add all inputs together.
+        for (int i = 0; i < inputs.length; i++) {
+            rawOutput += inputs[i] * weights[i];
+        }
+
+        //Bias neron influence
+        if (Hyperparameters.USE_BIAS_NEURONS)
+            rawOutput += weights[weights.length - 1];
+
+        return rawOutput;
+    }
+}
