--- src/resources/fxml/bindings/MainWindow_en.properties
@@ -1,7 +1,0 @@
-#Localization file for english language
-
-tab.data=Data
-tab.architecture=Architecture
-tab.management=Management
-tab.results=Results
-tab.about=About
