--- src/resources/fxml/bindings/MainWindow_ru.properties
@@ -1,7 +1,0 @@
-#Localization file for russian language
-
-tab.data=Данные
-tab.architecture=Архитектура
-tab.management=Управление
-tab.results=Результаты
-tab.about=О приложении
